# wwwdir/venom/shared/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  description: VENOM CODES API specification — WHMCS-backed client portal
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/invoices:
    get:
      tags: [Billing]
      summary: List invoices
      operationId: getInvoices
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
  /api/invoices/{invoiceId}:
    get:
      tags: [Billing]
      summary: Get invoice
      operationId: getInvoice
      security: [{ bearerAuth: [] }]
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
  /api/domains:
    get:
      tags: [Domains]
      summary: List domains
      operationId: getDomains
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/domains/whois:
    get:
      tags: [Domains]
      summary: WHOIS lookup
      operationId: whoisLookup
      parameters:
        - name: domain
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
  /api/knowledgebase/search:
    get:
      tags: [Public]
      summary: Search KB
      operationId: searchKnowledgebase
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KbSearchResponse"
...
  /api/invoices/{invoiceId}/pay:
    post:
      tags: [Billing]
      operationId: payInvoice
      security: [{ bearerAuth: [] }]
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
  /api/domains/{domainId}:
    get:
      tags: [Domains]
      operationId: getDomain
      security: [{ bearerAuth: [] }]
      parameters:
        - name: domainId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/domains/{domainId}/nameservers:
    put:
      tags: [Domains]
      operationId: updateDomainNameservers
      security: [{ bearerAuth: [] }]
      parameters:
        - name: domainId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [nameservers]
              properties:
                nameservers:
                  type: array
                  items:
                    type: string
                  minItems: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/domains/{domainId}/renew:
    post:
      tags: [Domains]
      operationId: renewDomain
      security: [{ bearerAuth: [] }]
      parameters:
        - name: domainId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                years:
                  type: integer
                  minimum: 1
                  default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/domains/{domainId}/autorenew:
    post:
      tags: [Domains]
      operationId: toggleDomainAutoRenew
      security: [{ bearerAuth: [] }]
      parameters:
        - name: domainId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/domains/{domainId}/idprotection:
    post:
      tags: [Domains]
      operationId: toggleDomainIdProtection
      security: [{ bearerAuth: [] }]
      parameters:
        - name: domainId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enable:
                  type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/domains/{domainId}/registrarlock:
    post:
      tags: [Domains]
      operationId: toggleDomainRegistrarLock
      security: [{ bearerAuth: [] }]
      parameters:
        - name: domainId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                lock:
                  type: boolean
                  default: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/domains/{domainId}/transfer:
    post:
      tags: [Domains]
      operationId: transferDomain
      security: [{ bearerAuth: [] }]
      parameters:
        - name: domainId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eppCode]
              properties:
                eppCode:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
  /api/knowledgebase/categories/{categoryId}:
    get:
      tags: [Public]
      operationId: getKbCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KbCategoryResponse"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/knowledgebase/categories/{categoryId}/articles/{articleId}:
    get:
      tags: [Public]
      operationId: getKbArticle
      parameters:
        - name: categoryId
          in: path
          required: true
          schema: { type: string }
        - name: articleId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KbArticleResponse"
        "404":
          description: Article not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
...
    AnyJson:
      type: object
      additionalProperties: true
...
    InvoiceListItem:
      type: object
      required: [id, status, total]
      properties:
        id:
          type: string
        status:
          type: string
        date:
          type: string
        dueDate:
          type: string
        total:
          type: string
        subtotal:
          type: string
        taxAmount:
          type: string
        credit:
          type: string
        paymentMethod:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    InvoiceListResponse:
      type: object
      required: [invoices, total]
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceListItem"
        total:
          type: number
...
    KbCategory:
      type: object
      required: [id, parentid, name, language]
      properties:
        id:
          type: integer
        parentid:
          type: integer
        name:
          type: string
        description:
          type: string
        language:
          type: string
...
    KbArticle:
      type: object
      required: [id, title, article, views, useful, votes, catid, language]
      properties:
        id:
          type: integer
        title:
          type: string
        article:
          type: string
        views:
          type: integer
        useful:
          type: integer
        votes:
          type: integer
        catid:
          type: integer
        parentid:
          type: integer
        language:
          type: string
...
    KbSearchResponse:
      type: object
      required: [articles]
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/KbArticle"
    KbCategoryResponse:
      type: object
      required: [category, articles]
      properties:
        category:
          $ref: "#/components/schemas/KbCategory"
        articles:
          type: array
          items:
            $ref: "#/components/schemas/KbArticle"
    KbArticleResponse:
      type: object
      required: [category, article]
      properties:
        category:
          $ref: "#/components/schemas/KbCategory"
        article:
          $ref: "#/components/schemas/KbArticle"